prefect aht10

+++ components/olcd/include/olcd.h
@@ -3,7 +3,7 @@
 #include "lvgl.h"
 typedef struct _olcd_data {
   lv_disp_t *disp;
-  int *data;
+  float *data;
 } olcd_data;
 lv_disp_t *setup_olcd(void);
 void run_olcd(void *);
